refactor: SummaryView를 ViewController에서 분리하고 delegate 패턴 적용
- summary UI 컴포넌트와 로직을 SummaryView로 분리
- summary 펼침/접힘 상태 변화를 delegate로 ViewController에 전달
- ViewController에서 summary 관련 프로퍼티 및 메서드 정리
- ViewController는 SummaryView 구성 및 상태 저장만 담당하도록 개선
